add listview for cities

+++ app/src/main/java/com/jurajsallai/superweatherapp/datamodel/Wind.kt
@@ -10,6 +10,6 @@ class Wind {
     var speed: Double? = null
     @SerializedName("deg")
     @Expose
-    var deg: Int? = null
+    var deg: Double? = null
 
 }
